Add can_assign and can_assign_add keys to the roles endpoint
This is to inform the front-end whether role assignment should be
allowed or not. It was necessary to create the can_assign_add key,
because a user can only create another with the roles they have. But in
editing he can assign other roles.
